requirements.txt
Fabric==1.8.0
PyYAML==3.10
South==0.8.2
argparse==1.2.1
boto==2.15.0
coverage==3.7.1
django-debug-toolbar==0.10.2
ecdsa==0.10
paramiko==1.12.0
psycopg2==2.4.6
pycrypto==2.6.1
python-coveralls==2.4.2
python-memcached==1.53
requests==2.1.0
sh==1.09
six==1.5.1
sqlparse==0.1.10
